+++ alacritty/src/config/ui_config.rs
@@ -69,7 +69,8 @@ pub struct UiConfig {
     mouse_bindings: MouseBindings,
 
     /// Background opacity from 0.0 to 1.0.
-    background_opacity: Percentage,
+    #[config(deprecated = "use window.opacity instead")]
+    window_opacity: Option<Percentage>,
 }
 
 impl Default for UiConfig {
@@ -84,7 +85,7 @@ impl Default for UiConfig {
             config_paths: Default::default(),
             key_bindings: Default::default(),
             mouse_bindings: Default::default(),
-            background_opacity: Default::default(),
+            window_opacity: Default::default(),
             bell: Default::default(),
             colors: Default::default(),
             draw_bold_text_with_bright_colors: Default::default(),
@@ -115,8 +116,8 @@ impl UiConfig {
     }
 
     #[inline]
-    pub fn background_opacity(&self) -> f32 {
-        self.background_opacity.as_f32()
+    pub fn window_opacity(&self) -> f32 {
+        self.window_opacity.unwrap_or(self.window.opacity).as_f32()
     }
 
     #[inline]